requirements.txt
pytest
pytest_bdd
selenium==3.141.0
urllib3==1.26.16
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest
robotframework